Moves CartContext to new folder
